dozeアラートをメイン画面に移動
